// src/shared/configuration/configuration.enum.ts
/* eslint-disable prettier/prettier */
export enum Configuration {
    HOST = 'HOST',
    PORT = 'PORT',
    HOST_POSTGRES = 'HOST_POSTGRES',
    PORT_POSTGRES = 'PORT_POSTGRES',
    USER_POSTGRES = 'USER_POSTGRES',
    PASSWORD_POSTGRES = 'PASSWORD_POSTGRES',
    DATABASE_POSTGRES = 'DATABASE_POSTGRES',
    SCHEMA_POSTGRES = 'SCHEMA_POSTGRES',
    SYNCHRONIZE_POSTGRES = 'SYNCHRONIZE_POSTGRES',
    REDIS_CLIENT = 'REDIS_CLIENT',
    REDIS_PORT = 'REDIS_PORT',
    REDIS_USER = 'REDIS_USER',
    REDIS_PASSWORD = 'REDIS_PASSWORD',
    JWT_CUSTOMER_ACCESS_SECRET_REFRESH = 'JWT_CUSTOMER_ACCESS_SECRET_REFRESH',
    JWT_CUSTOMER_ACCESS_SECRET = 'JWT_CUSTOMER_ACCESS_SECRET',
    JWT_CUSTOMER_ACCESS_EXPIRATION_TIME = 'JWT_CUSTOMER_ACCESS_EXPIRATION_TIME',
    JWT_CUSTOMER_TYPE_SECRET_APP = 'JWT_CUSTOMER_TYPE_SECRET_APP',
    JWT_CUSTOMER_CLIENT_HOST_APP = 'JWT_CUSTOMER_CLIENT_HOST_APP',
    JWT_ADMIN_ACCESS_SECRET_REFRESH = 'JWT_ADMIN_ACCESS_SECRET_REFRESH',
    JWT_ADMIN_ACCESS_SECRET = 'JWT_ADMIN_ACCESS_SECRET',
    JWT_ADMIN_ACCESS_EXPIRATION_TIME = 'JWT_ADMIN_ACCESS_EXPIRATION_TIME',
    JWT_ADMIN_TYPE_SECRET_APP = 'JWT_ADMIN_TYPE_SECRET_APP',
    JWT_ADMIN_CLIENT_HOST_APP = 'JWT_ADMIN_CLIENT_HOST_APP',
    SETTING_ADMIN_SECRET_KEY = 'SETTING_ADMIN_SECRET_KEY',
}